decode.cpp fixed bug in timestamp decoding
Higher words needed to be converted to 64-bit explicitly before bit-shifting. Tested and works now.
Not sure if explicit conversion is necessary for 12 bit shift but did anyway for both timestamp and event number

+++ decode/decode.cpp
@@ -145,13 +145,13 @@ int main( int argc, const char **argv) {
           fine_timestamp = get_fine_timestamp( data );
         }
         else if( iFeuH == 4 ){
-          eventID  += get_Event_ID( data )<<12;
+          eventID  += (uint64_t)get_Event_ID( data )<<12;
         }
         else if( iFeuH == 5 ){
-          timestamp +=  get_timestamp( data )<<12;
+          timestamp +=  (uint64_t)get_timestamp( data )<<12;
         }
         else if( iFeuH == 6 ){
-          timestamp +=  get_timestamp( data )<<24;
+          timestamp +=  (uint64_t)get_timestamp( data )<<24;
         }
         else if( iFeuH == 7 ){
           timestamp +=  (uint64_t)get_timestamp( data )<<36;
